Fix data type in for loop and Fix coding standard

+++ Lab8/FixedVector.h
@@ -97,7 +97,7 @@ namespace lab8
 	template<class T, size_t N>
 	int FixedVector<T, N>::GetIndex(const T& t) const
 	{
-		for (int i = 0; i < mSize; ++i)
+		for (size_t i = 0; i < mSize; ++i)
 		{
 			if (mMainArray[i] == t)
 			{
